requirements.txt
beautifulsoup4==4.7.1
certifi==2019.3.9
chardet==3.0.4
Click==7.0
idna==2.8
motionless==1.3.2
python-craigslist==1.0.6
python-dateutil==2.8.0
pytz==2018.9
PyYAML==5.1
requests==2.21.0
six==1.12.0
slackclient==1.3.1
soupsieve==1.9
urllib3==1.24.1
websocket-client==0.54.0
